bid: make work on x86
bid is a software-only implementation of the dfp standard. Don't scare
away x86 users by intercepting nearly all of the toolchain provided
implementations we already assume are provided by gcc.

This has the added benefit of making long double work correctly
for x86. Given the brokenness, it is probably safe to assume
there are no users on x86.

Instead, provide a good experience for the libm-like ABI defined
by TS 18661-2 which are not supplied by the toolchain.

Note, there is one exception to the above, __int128 provided by
GCC. libgcc does not provide conversion routines for this type.

Likewise, TImode routines seem to be missing for _Decimal{32,64}.