isunordered*: use GCC builtin
This was discovered after the s390 variant generated warnings.

Use the GCC builtin to implement this function as it sees fit for
all targets. Likewise, remove the now superfluous s390 variants.

This has the undersirable side-effect of calling out to unord*d2
on soft-dfp platforms. Thus, PLT bypass support is added for the
unordered GCC builtin functions too.