gen-auto-dfp-tests: improve
Using a single high precision value and approximating the
input and outputs to a _DecimalN -> _DecimalN function is a
bad idea.

This is easily demonstrated via the exp() test which today
just calls libdecnumber implementation which is correct to 1ULP,
and likewise checked against my TI-30XIIS.

For now, roundtrip the numbers against the precision of the
format to better approximate the decimal value under test.

Likewise, sample test inputs are taken from glibc, and modified
to run against DFP. Argubly specifying tests in BFP makes no
sense for DFP tests, so we change accordingly.